Added NxN grid and ObstacleGrid set_obstacle
To modify the tank game code to allow square obstacles to be added to the screen using a mouse click and store them in an NxN grid, we can start by creating a new class called ObstacleGrid. In this class, we will define the data structure to store the obstacles in an NxN grid, starting with no obstacles.